active_activity_session#update relax params (#8870)
* relaxing params

* file rename

* solution

* another spec

* feedback
